ZipDirectoryOperation fixes. Zipped folder won't retain the full directory structure starting at root level. Status updater will show the file name. Compression ratio can't be below 0 -> NSUInteger type.

+++ SpriteBuilder/ccBuilder/ZipDirectoryOperation.h
@@ -5,7 +5,6 @@
 
 @property (nonatomic, copy) NSString *inputPath;
 @property (nonatomic, copy) NSString *zipOutputPath;
-
-@property (nonatomic) int compression;
+@property (nonatomic) NSUInteger compression;
 
 @end
